setup/requirements.txt
# requirements.txt
# python_version == 3.11.9 commented out but YOU STILL NEED THIS
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
edge-tts==7.0.2
executing==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
h3==4.2.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
omegaconf==2.3.0
openai==1.77.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.7
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
Resemblyzer==0.1.4
scikit-learn==1.6.1
scipy==1.15.3
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.2
srt==3.5.3
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
timezonefinder==6.5.9
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webrtcvad==2.0.10
yarl==1.20.0